[fix]: reorganized src folder
