+++ .gitignore
@@ -101,3 +101,4 @@ paper/arxiv_submission_v3.tar.gz
 paper/arxiv_submission_v4.tar.gz
 
 *.zip
+v2_consolidated/
