+++ src/visualization/regenerate_figure1_all_models.py
