+++ src/visualization/regenerate_figure2_all_models.py
