+++ src/visualization/regenerate_figure3.py
